requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
nltk>=3.8.1
torch>=2.0.0
praw>=7.7.1
tqdm>=4.66.0
wordcloud>=1.9.3
streamlit-aggrid
python-dotenv>=1.0.0
scipy>=1.10.0
protobuf>=3.20.0
loguru>=0.7.2